% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{convertR}
\alias{convertR}
\title{Convert TQ .raw to .mzml}
\usage{
convertR(
  filename = NULL,
  outpath = NULL,
  msconvert = NULL,
  args = c("--mzML", "--64", "--zlib", "--chromatogramFilter \\"index [1,100]\\"")
)
}
\arguments{
\item{filename}{Path to the desired .raw folder}

\item{outpath}{Path to export the .mzml file to}

\item{msconvert}{Path to your proteowizard msconvert.exe}

\item{args}{Additional command line arguments, fille by default}
}
\value{
Returns nothing in the session, .mzml will be exported to given folder
}
\description{
Convert TQ .raw to .mzml
}
\examples{
msconexe <- "path/to/msconvert.exe"
convertR(filename = "C:/some_tq_data.raw", outpath = "C:/my_tq_data", msconvert = msconexe)
}
