% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{estDims}
\alias{estDims}
\title{Dimension estimation 1}
\usage{
estDims(filename = NULL)
}
\arguments{
\item{filename}{Path to the desired .raw folder}
}
\value{
The x,y dimension lengths
}
\description{
Estimate pixel dimensions from an _extern.inf file
}
\examples{
xydims <- estDims("C:/some_tq_data.raw");
}
