% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{estDims2}
\alias{estDims2}
\title{Dimension estimation 2}
\usage{
estDims2(n)
}
\arguments{
\item{n}{The total number of scans in a file}
}
\value{
A list of potential pixel dimensions for the given length
}
\description{
Estimate pixel dimensions from from total scan number
}
\examples{
dimslist <- estDims2(40000);
}
