% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{getData}
\alias{getData}
\title{Get a data matrix from SRM data}
\usage{
getData(srm_in, addpix = TRUE)
}
\arguments{
\item{srm_in}{Input data from an MSNbase SRM object}

\item{addpix}{Whether to make up for a "missing pixel" or not, defaults to TRUE}
}
\value{
Matrix of TQ data, scans in rows, ions in columns
}
\description{
Get a data matrix from SRM data
}
\examples{
data <- MSnbase::readSRMData("C:/my_tq_data/my_file.mzML")
datamatrix <- getData(data)
}
