% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{getNames}
\alias{getNames}
\title{Get names of ion channels from acquisition file}
\usage{
getNames(filename = NULL)
}
\arguments{
\item{filename}{Path to the desired .raw folder}
}
\value{
Array of strings of >2 characters in length
}
\description{
Get names of ion channels from acquisition file
}
\examples{
compund_names <- getNames("C:/some_tq_data.raw")
}
