% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{getPos}
\alias{getPos}
\title{Dimension estimation 3}
\usage{
getPos(filename = NULL, exe = "MRM")
}
\arguments{
\item{filename}{Path to the desired .raw folder}

\item{exe}{Path to Waters-provided .exe file, defaults to TQ specific MRMProcessing.exe}
}
\value{
A data frame of x,y coordinate pairs for each scan
}
\description{
Get global coordinate positions reported from the instrument/stage
}
\examples{
xypos <- getPos("C:/some_tq_data.raw")
xydims <- c(length(unique(xypos[,1], unique(xypos[,2]))))
}
