% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{plotDims}
\alias{plotDims}
\title{Dimension pair plotting}
\usage{
plotDims(dims_list, input_data)
}
\arguments{
\item{dims_list}{A list of coordinate pairs generated by estDims2}

\item{input_data}{A data matrix generated from data import functions}
}
\value{
A data frame of x,y coordinate pairs for each scan
}
\description{
Plot every possible coordinate pair from an estDims2 generated list
}
\examples{
plotlist <- plotDims(dimslist, datamatrix)
plotlist[[10]]
}
