% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_tools.R
\name{roiSelect}
\alias{roiSelect}
\title{Get region of interest using point selection}
\usage{
roiSelect(
  image,
  dims,
  log = FALSE,
  flipy = FALSE,
  invert = FALSE,
  compound = "tic"
)
}
\arguments{
\item{image}{Data frame of your imaging data, scans as rows}

\item{dims}{x,y dimension pair of your pixel dimensions}

\item{log}{Whether to log transform your image, defaults to FALSE}

\item{flipy}{Whether to mirror the y axis, defaults to FALSE}

\item{invert}{Whether to invert pixel values, defaults to false}

\item{compound}{Which column to visualise for ROI selection, defaults to "tic" (sum of all)}
}
\value{
Numeric vector of scan numbers in ROI
}
\description{
Get region of interest using point selection
}
